fix(addons): address review findings (MINI-1)
- Forward TailscaleService.validate().errorCode through the settings POST
  response as validationErrorCode so the client can switch on the
  structured TailscaleErrorCode instead of substring-matching English
  validation messages (server/src/routes/tailscale-settings.ts,
  client/src/app/connectivity/tailscale/page.tsx, lib/types/tailscale.ts).
- Consolidate buildAclSnippet into @mini-infra/types as the single source
  of truth; delete the byte-identical client and server copies; update the
  client unit test to import from the shared package.
